requirements.txt
python-binance==1.0.16
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.2
ta>=0.7.0
websockets==10.4
openai>=1.0.0
redis>=5.0.0
aioredis>=2.0.0
asyncio>=3.4.3
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
ujson>=5.8.0
